cloud-project/video-to-audio/services/converter_service/requirements.txt
pika==1.3.2
pymongo==4.6.0
psycopg2-binary==2.9.9

